requirements.txt
poetry
crewai
crewai-tools
openai
python-dotenv
streamlit
google-generativeai